Fixes content proxy performance on selects
Before, the content proxy would re-sort and re-filter
its content after any change to its content's labels
(e.g. if something is added/deleted to the content, or
if a label changed). This ultimately lead to O(n^2 * log(n))
performance for adding n items. This change makes it so that
the filteredContent and sortedFilteredContent properties are
updated in a debounced way which means that we should basically
have O(n * log(n)) performance (only the last sort matters).