Add `selectLastElement` to put the caret in the last paragraph.
Also adds a hack to allow doing this. For a still unknown reason,
placing the caret in an empty div element works in the library but not
in my subclass.

Based on
http://stackoverflow.com/questions/5488809/how-to-place-caret-inside-an-empty-dom-element-node
this is never supposed to work with webkit.

So this is a workaround adding an invisible character to the element.
This will need to be revisited at some point.