Fix popover teardown in popover links
When a popover link is destroyed, it should destroy the popover it has created,
if any. The code in place to do this didn't work because it was trying to hit
the computed property cache but the popover wasn't stored in a computed
property. The fix is to just get the popover and destroy it if it exists.