Prevent page refresh with color picker custom selection.
When using the color picker tool, users can enter a custom
hex value in an input field. There are two ways to trigger
the selection - (1) clicking the sample color tile to the
left of the input field, or (2) submitting the form by
pressing enter. Previously, submitting the form by
pressing enter resulted in a page refresh and the color
not actually being updated. This is because the default
form submission was not being overridden and the
`setCustomColor` action on insert-newline was never
actually getting triggered. Now, the action triggers
on form submit and the form works as expected.

Closes #236