Don't create select rows for empty grouping names (#270)
Currently there's no way to have some rows in a select dropdown show up under groupings, but others don't. If an option has an `optionGroupPath` of an empty string, it results in a bunch of whitespace in the dropdown. This PR just makes it so grouping rows with an empty name are ignored, which leads to the height of the dropdown calculating correctly, and no weird whitespace appearing.