Fix a race condition between value/content
Previously if a value argument was set before content had been set
(which may happen during initialization) then a selection value would be
set matching the value. This selection would probably be wrong,
especially if `optionValuePath` is also present, meaning selection is
expected to be an object.

This refactors some of the current implementation to make clearer what
the code intent is, and makes these changes:

* Rename setDefaultSelection to _updateSelectionFromContent, and include
  logic that skips the work during initialization and propagated the
  value -> selection if needed.
* Move _updateSelectionFromContent into `init` instead of
  `didInsertElement`. This work just needs to happen after arguments, not
  after rendering itself.
* Have the content.[] observer schedule into actions and minimize
  runs based on unsettled content.
* Add tests that show value -> selection and selection -> value being
  honored when the content property changes.