Add collapsible grouped headers to select-component (#288)
Add ability to collapse/expand group headers in `select-component` and `multi-select-component`. `optionGroupPath` should be set and `isGroupHeaderCollapsible=true` on the select.

There were 2 possible approaches to implementing the hiding of the list items after the header is collapsed
1. Use CSS `display: none` to hide the list item if the header is collapsed
2. Remove the list item from the backing JavaScript data structure for the select list

Approach 1 was chosen was approach 2 was not feasible. `vertical-collection` is used to render the list items to utilize its occlusion performance capability. There currently is not a built-in mechanism in `vertical-collection` to schedule a re-measure or a re-render.

![collapsible-group-headers](https://user-images.githubusercontent.com/42778466/54151395-1c36d880-4411-11e9-8aa8-59bde50b4f6e.gif)